Login Error Messages
Added In-Game Login and Registration error messages. No longer just in the Console log.
