+++ AR Collaboration/Assembly-CSharp.csproj
@@ -63,9 +63,9 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Assets\Login.cs" />
-    <Compile Include="Assets\MyActionScript.cs" />
-    <Compile Include="Assets\Register.cs" />
+    <Compile Include="Assets\Scripts\Login.cs" />
+    <Compile Include="Assets\Scripts\MyActionScript.cs" />
+    <Compile Include="Assets\Scripts\Register.cs" />
     <Compile Include="Assets\ZED\Examples\Dark Room\Scripts\DisableUntilZEDReady.cs" />
     <Compile Include="Assets\ZED\Examples\Dark Room\Scripts\LightShow.cs" />
     <Compile Include="Assets\ZED\Examples\Dark Room\Scripts\PitchOscillator.cs" />
