--- AR Collaboration/Assets/Login.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: cc57716cfe7f7b64ab33fb1fdb435ef4
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
