--- AR Collaboration/Assets/MyActionScript.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 7ac8013b2c93c844f86c28c07b2fa9ef
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
