--- AR Collaboration/Assets/Register.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6e0aa5c377c8f054a86a40bd42a91a9a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
