+++ AR Collaboration/Assets/Scripts/Login.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: cc57716cfe7f7b64ab33fb1fdb435ef4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
