+++ AR Collaboration/Library/UIElements/EditorWindows/UnityEditor.UIElements.Debugger.UIElementsDebugger.pref
@@ -1,0 +1,24 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &1
+MonoBehaviour:
+  m_ObjectHideFlags: 61
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 12386, guid: 0000000000000000e000000000000000, type: 0}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  m_Keys:
+  - unity-uie-debugger-splitter__UnityEditor.UIElements.Debugger.DebuggerSplitter
+  - __PanelContainer__rootVisualContainer__unity-uie-debugger-splitter__unity-tree-view__list-view__UnityEngine.UIElements.ListView
+  - __PanelContainer__rootVisualContainer__unity-uie-debugger-splitter__unity-tree-view__list-view__list-view__scroll-view__HorizontalScroller__Slider__UnityEngine.UIElements.Slider
+  - __PanelContainer__rootVisualContainer__unity-uie-debugger-splitter__unity-tree-view__list-view__list-view__scroll-view__VerticalScroller__Slider__UnityEngine.UIElements.Slider
+  m_Values:
+  - '{"leftPaneWidth":400.0}'
+  - '{"m_ScrollOffset":0.0,"m_SelectedIds":[]}'
+  - '{"m_Value":0.0}'
+  - '{"m_Value":0.0}'
